+++ tests/test_memo.py
@@ -2,8 +2,8 @@ import unittest
 import os
 from pprint import pprint
 from itertools import cycle
-from bitsharesbase.account import BrainKey, Address, PublicKey, PrivateKey
-from bitsharesbase.memo import (
+from graphsdkbase.account import BrainKey, Address, PublicKey, PrivateKey
+from graphsdkbase.memo import (
     get_shared_secret,
     _pad,
     _unpad,